+++ src/main/kotlin/controllers/PrinterUIController.kt
@@ -12,7 +12,7 @@ class PrinterUIController : Controller() {
     var printers = PrinterJSONStore()
     var logger = KotlinLogging.logger {}
 
-    fun add(_name: String, _price : Double, _watts : Int, _investment : Int) {
+    fun add(_name: String, _price : Float, _watts : Int, _investment : Int) {
         var aPrinter = PrinterModel(printerName = _name, printerPrice = _price, wattUsage = _watts, investmentReturn = _investment)
         printers.create(aPrinter)
         logger.info("Printer Added")
